backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
requests
httpx
pydantic
typing-extensions
